write: Test Append
