+++ go.mod
@@ -2,4 +2,4 @@ module neilpa.me/go-jfif
 
 go 1.14
 
-require neilpa.me/go-x v0.2.1-0.20200507232743-5243b9624d5e
+require neilpa.me/go-x v0.2.1-0.20200508010119-d3561a1b764a
